fix: rename githost form
