+++ package.json
@@ -11,7 +11,7 @@
 		"dev:ui": "pnpm --filter @gitbutler/ui storybook",
 		"dev:web": "turbo run --filter @gitbutler/web dev --no-daemon",
 		"dev:desktop": "pnpm tauri dev",
-		"dev:internal-tauri": "turbo run --filter @gitbutler/desktop dev",
+		"dev:internal-tauri": "turbo watch --filter @gitbutler/desktop dev",
 		"test": "turbo run test --no-daemon",
 		"test:watch": "pnpm --filter @gitbutler/desktop run test:watch",
 		"test:e2e": "pnpm --filter @gitbutler/desktop run test:e2e",
